refactor : 웹소켓 관련 파일 이동
